# contract/openapi.yaml
openapi: 3.0.0
info:
  title: Kafbat UI Api Service
  contact:
    name: API Support
    email: support@kafbat.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.2.0
...
paths:
...
  /api/clusters/{clusterName}/ksql/response:
    get:
      operationId: openKsqlResponsePipe
      summary: openKsqlResponsePipe
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
        - name: pipeId
          in: query
          required: true
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            text/event-stream:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KsqlResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiNotFoundResponse'
      tags:
        - Ksql
  /api/clusters/{clusterName}/ksql/streams:
    get:
      operationId: listStreams
      summary: listStreams
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KsqlStreamDescription'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiNotFoundResponse'
      tags:
        - Ksql
  /api/clusters/{clusterName}/ksql/tables:
    get:
      operationId: listTables
      summary: listTables
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KsqlTableDescription'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiNotFoundResponse'
      tags:
        - Ksql
  /api/clusters/{clusterName}/ksql/v2:
    post:
      operationId: executeKsql
      summary: executeKsql
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KsqlCommandV2Response'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiNotFoundResponse'
      tags:
        - Ksql
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KsqlCommandV2'
...
components:
  schemas:
...
    ApiNotFoundResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      description: ''
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
        - requestId
      properties:
        code:
          type: integer
          format: int32
          description: Internal error code (can be used for message formatting & localization on UI)
        message:
          type: string
          description: Error message
        timestamp:
          type: number
          format: decimal
          description: Response unix timestamp in ms
        requestId:
          type: string
          description: Unique server-defined request id for convenient debugging
        fieldsErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        stackTrace:
          type: string
    FieldError:
      type: object
      required:
        - fieldName
      properties:
        fieldName:
          type: string
          description: Name of field that violated format
        restrictions:
          type: array
          items:
            type: string
          description: Field format violations description (ex. ["size must be between 0 and 20", "must be a well-formed email address"])
...
    KsqlCommandV2:
      type: object
      required:
        - ksql
      properties:
        ksql:
          type: string
        streamsProperties:
          type: object
          additionalProperties:
            type: string
    KsqlCommandV2Response:
      type: object
      required:
        - pipeId
      properties:
        pipeId:
          type: string
    KsqlResponse:
      type: object
      properties:
        table:
          $ref: '#/components/schemas/KsqlTableResponse'
    KsqlStreamDescription:
      type: object
      properties:
        name:
          type: string
        topic:
          type: string
        keyFormat:
          type: string
        valueFormat:
          type: string
    KsqlTableDescription:
      type: object
      properties:
        name:
          type: string
        topic:
          type: string
        keyFormat:
          type: string
        valueFormat:
          type: string
        isWindowed:
          type: boolean
    KsqlTableResponse:
      type: object
      properties:
        header:
          type: string
        columnNames:
          type: array
          items:
            type: string
        values:
          type: array
          items:
            type: array
            items: {}
...
servers:
  - url: http://localhost:8080
    description: Default endpoint for Kafbat UI API
    variables: {}